move notebook related funcs to bokeh.io.notebook
